+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -19,10 +19,13 @@ exclude: true
   ([bug 1836356]({{bugzilla}}1836356))
 - Added [`PromptInstanceDelegate.onSelectIdentityCredentialAccount`][116.3] to allow the user to choose an account on the Identity Credential Provider (FedCM) they previously chose to be used when authenticating.
     ([bug 1836363]({{bugzilla}}1836363))
-
+- Added [`PromptInstanceDelegate.onShowPrivacyPolicyIdentityCredential`][116.4] to allow the user to indicate if agrees or not with the privacy policy of the Identity Credential provider.
+  ([bug 1836358]({{bugzilla}}1836358))
+-
 [116.1]: {{javadoc_uri}}/GeckoSession.html#didPrintPageContent
 [116.2]:{{javadoc_uri}}/GeckoSession.PromptDelegate.html#onSelectIdentityCredentialProvider(org.mozilla.geckoview.GeckoSession,org.mozilla.geckoview.GeckoSession.PromptDelegate.IdentityCredential.ProviderSelectorPrompt)
 [116.3]:{{javadoc_uri}}/GeckoSession.PromptDelegate.html#onSelectIdentityCredentialAccount(org.mozilla.geckoview.GeckoSession,org.mozilla.geckoview.GeckoSession.PromptDelegate.IdentityCredential.AccountSelectorPrompt)
+[116.4]:{{javadoc_uri}}/GeckoSession.PromptDelegate.html#onShowPrivacyPolicyIdentityCredential(org.mozilla.geckoview.GeckoSession,org.mozilla.geckoview.GeckoSession.PromptDelegate.IdentityCredential.PrivacyPolicyPrompt)
 
 ## v115
 - Changed [`SessionPdfFileSaver.createResponse`][115.1] to response of saving PDF to accept two additional
@@ -1387,4 +1390,4 @@ to allow adding gecko profiler markers.
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport(android.content.Context,android.os.Bundle,java.lang.String)
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: 5faf2e6a215dfd661f1f15a9e846ca0a5c835f18
+[api-version]: 7f26c6109e435702d0623f02592ae21d8c721f95